Move existing providers to folder
